Create tests module and split tests.py
