+++ marcador/tests/__init__.py
@@ -1,0 +1,8 @@
+from .models import TagTestCase, BookmarkTestCase
+from .views import (
+    BookmarkListTestCase,
+    UserBookmarkListTestCase,
+    BookmarkCreateTestCase,
+    BookmarkUpdateTestCase,
+    BookmarkDeleteTestCase
+)
